requirements.txt
pandas==2.2.2
numpy==2.0.1
python-dotenv==1.0.1
MetaTrader5==5.0.45
flask==3.0.3
pytz==2024.1
openai==1.43.0
openpyxl==3.1.5
matplotlib==3.9.0
requests==2.32.3
pyttsx3==2.90
